gltf: Move file I/O functionality to fileio.cpp
Also add forgotten comment to basistoktx.cpp
