+++ gltf/gltfpack.cpp
@@ -703,98 +703,6 @@ std::string inferMimeType(const char* path)
 		return "image/" + extl;
 }
 
-std::string getFullPath(const char* path, const char* base_path)
-{
-	std::string result = base_path;
-
-	std::string::size_type slash = result.find_last_of("/\\");
-	result.erase(slash == std::string::npos ? 0 : slash + 1);
-
-	result += path;
-
-	return result;
-}
-
-std::string getFileName(const char* path)
-{
-	std::string result = path;
-
-	std::string::size_type slash = result.find_last_of("/\\");
-	if (slash != std::string::npos)
-		result.erase(0, slash + 1);
-
-	std::string::size_type dot = result.find_last_of('.');
-	if (dot != std::string::npos)
-		result.erase(dot);
-
-	return result;
-}
-
-bool readFile(const char* path, std::string& data)
-{
-	FILE* file = fopen(path, "rb");
-	if (!file)
-		return false;
-
-	fseek(file, 0, SEEK_END);
-	long length = ftell(file);
-	fseek(file, 0, SEEK_SET);
-
-	if (length <= 0)
-	{
-		fclose(file);
-		return false;
-	}
-
-	data.resize(length);
-	size_t result = fread(&data[0], 1, data.size(), file);
-	fclose(file);
-
-	return result == data.size();
-}
-
-bool writeFile(const char* path, const std::string& data)
-{
-	FILE* file = fopen(path, "wb");
-	if (!file)
-		return false;
-
-	size_t result = fwrite(&data[0], 1, data.size(), file);
-	fclose(file);
-
-	return result == data.size();
-}
-
-struct TempFile
-{
-	std::string path;
-	int fd;
-
-	TempFile(const char* suffix)
-	    : fd(-1)
-	{
-#ifdef _WIN32
-		const char* temp_dir = getenv("TEMP");
-		path = temp_dir ? temp_dir : ".";
-		path += "\\gltfpack-XXXXXX";
-		(void)_mktemp(&path[0]);
-		path += suffix;
-#else
-		path = "/tmp/gltfpack-XXXXXX";
-		path += suffix;
-		fd = mkstemps(&path[0], strlen(suffix));
-#endif
-	}
-
-	~TempFile()
-	{
-		unlink(path.c_str());
-#ifndef _WIN32
-		close(fd);
-#endif
-	}
-};
-
 bool checkBasis()
 {
 	const char* basisu_path = getenv("BASISU_PATH");
